refactor(tools): split grok-native into focused modules
Extract workspace path containment, local grep, and read/search-replace
executors from grok-native.ts into grok-native-paths.ts,
grok-native-grep.ts, and grok-native-file-tools.ts. Keep grok-native.ts
focused on tool registration, schemas, sync, and web search orchestration.

Behavior unchanged: all grok-native test suites and full npm test pass.
